requirements.txt
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.0
flask-jwt-extended
